# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/rfq_items:
    get:
      tags:
        - RfqItems
      summary: List RfqItems
      description: Get a paginated list of RfqItems with optional filtering and sorting
      operationId: listRfqItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - RfqItems
      summary: Create RfqItem
      description: Create a new RfqItem
      operationId: createRfqItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRfqItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/rfq_suppliers:
    get:
      tags:
        - RfqSuppliers
      summary: List RfqSuppliers
      description: Get a paginated list of RfqSuppliers with optional filtering and sorting
      operationId: listRfqSuppliers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqSupplierList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - RfqSuppliers
      summary: Create RfqSupplier
      description: Create a new RfqSupplier
      operationId: createRfqSupplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRfqSupplierRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfqSupplier'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    RfqItem:
      type: object
      required:
        - id
        - rfq_id
        - company_id
        - item_id
        - quantity
        - metadata
      properties:
        id:
          type: string
          format: uuid
        rfq_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        quantity:
          type: string
        metadata:
          type: object
...
    CreateRfqItemRequest:
      type: object
      required:
        - rfq_id
        - company_id
        - item_id
        - quantity
        - metadata
      properties:
        rfq_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        quantity:
          type: string
        metadata:
          type: object
...
    RfqItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RfqItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    RfqSupplier:
      type: object
      required:
        - id
        - rfq_id
        - company_id
        - supplier_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        rfq_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateRfqSupplierRequest:
      type: object
      required:
        - rfq_id
        - company_id
        - supplier_id
        - metadata
      properties:
        rfq_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    RfqSupplierList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RfqSupplier'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []